+++ src/Path/AliasRoutingPath.php
@@ -2,26 +2,32 @@
 namespace Pyncer\Routing\Path;
 
 use Pyncer\Routing\Path\AbstractRoutingPath;
-
-use function preg_match;
+use Pyncer\Validation\Rule\AliasRule;
+use Pyncer\Validation\ValueValidator;
 
 class AliasRoutingPath extends AbstractRoutingPath
 {
+    protected ValueValidator $validator;
+
     public function __construct(
         ?string $queryName = 'alias',
         ?string $routeDirPath = '@alias'
     ) {
         parent::__construct($queryName, $routeDirPath);
+
+        $this->validator = new ValueValidator();
+        $this->validator->AddRules(
+            new AliasRule(
+                allowNumericCharacters: true,
+                allowLowerCaseCharacters: true,
+                allowUpperCaseCharacters: true,
+                separatorCharacters: '-_',
+            )
+        );
     }
 
     public function isValidPath(string $path): bool
     {
-         $pattern = '/^[0-9a-z-_]+$/';
-
-         if (preg_match($pattern, $path)) {
-            return true;
-         }
-
-         return false;
+        return $this->validator->isValidAndClean($path);
     }
 }