+++ src/Path/IdRoutingPath.php
@@ -25,6 +25,6 @@ class IdRoutingPath extends AbstractRoutingPath
 
     public function isValidPath(string $path): bool
     {
-        return $this->validator->isValid($path);
+        return $this->validator->isValidAndClean(intval($path));
     }
 }