+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.1 - 2023-09-05
+
+### Changed
+
+- Alias and glob routing paths now use alias validation rule.
+- Path validation more strict.
+
 ## 1.1.0 - 2023-01-14
 
 ### Fixed